appMedicBack/requirements.txt
Authlib==1.3.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyasn1==0.6.0
pycparser==2.22
pymongo==4.7.2
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
Werkzeug==3.0.3
